// src/models/IRawInputRow.ts
export default interface IRawInputRow {
    "date": string;
    "note": string;
    "bodyWeightKg": string;
    "bodyWeightLb": string;
    "ex1_Name": string;
    "ex1_WeightKg": string;
    "ex1_WeightLb": string;
    "ex1_Set1": string;
    "ex1_Set2": string;
    "ex1_Set3": string;
    "ex1_Set4": string;
    "ex1_Set5": string;
    "ex2_Name": string;
    "ex2_WeightKg": string;
    "ex2_WeightLb": string;
    "ex2_Set1": string;
    "ex2_Set2": string;
    "ex2_Set3": string;
    "ex2_Set4": string;
    "ex2_Set5": string;
    "ex3_Name": string;
    "ex3_WeightKg": string;
    "ex3_WeightLb": string;
    "ex3_Set1": string;
    "ex3_Set2": string;
    "ex3_Set3": string;
    "ex3_Set4": string;
    "ex3_Set5": string;
    "ex4_Name": string;
    "ex4_WeightKg": string;
    "ex4_WeightLb": string;
    "ex4_Set1": string;
    "ex4_Set2": string;
    "ex4_Set3": string;
    "ex4_Set4": string;
    "ex4_Set5": string;
}